pyproject.toml
[tool.poetry]
name = "tweets_to_character"
version = "0.1.0"
description = ""
authors = ["Mohammed Arshath <marshath@gmail.com>"]
readme = "README.md"
include = ["tweets_to_character"]

[tool.poetry.dependencies]
python = ">=3.10,<4.0"
instructor = "^1.6.4"
openai = "^1.54.4"
pydantic = "^2.9.2"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"